add cors dependency and move user route to api folder
