+++ server/app.js
@@ -4,9 +4,10 @@ const connectDB = require('./config/db');
 const path = require('path');
 const cookieParser = require('cookie-parser');
 const morgan = require('morgan');
+const cors = require('cors');
 
 const indexRouter = require('./routes/index');
-const usersRouter = require('./routes/users');
+const usersRouter = require('./routes/api/users');
 
 const app = express();
 
@@ -21,6 +22,7 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
+app.use(cors({require: true, credentials: true}));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
