+++ server/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "config": "^3.3.6",
     "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
     "debug": "~2.6.9",
     "express": "^4.16.4",
     "http-errors": "~1.6.3",
